Maryia/moved stories related files to stories dir
